Separate classes for chat manager and store
