+++ tslint.json
@@ -5,7 +5,6 @@
     "no-duplicate-variable": true,
     "curly": true,
     "class-name": true,
-    "semicolon": [true, "always"],
     "triple-equals": true
   },
   "defaultSeverity": "warning"
